+++ src/discopop_webview_provider.ts
@@ -1,24 +1,33 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import * as vscode from 'vscode';
 import { getFiles } from './misc/iomanip';
+import { SourceHighlighting } from './misc/source_highlighting';
 
 
 import { executeCUGenTask } from './tasks/task_cu_gen';
 import { executeDepProfTask } from './tasks/task_dep_prof';
 import { executeFileMappingTask } from './tasks/task_file_mapping';
+import { executePatIdTask } from './tasks/task_pat_id';
 import { executeRedOpTask } from './tasks/task_red_op';
 
 export class DiscoPoPViewProvider implements vscode.WebviewViewProvider {
   folderPath: string;
   buildPath: string;
+  discopopPath: string;
 
   init: boolean;
   webview: vscode.WebviewView | undefined;
   interval!: NodeJS.Timeout;
 
-  taskMap : Record<string, Function>;
+  lastFiles: string[] = [];
 
-  constructor() {
+  sourceHighlighting: SourceHighlighting;
+
+  taskMap: Record<string, Function>;
+
+  constructor(sourceHighlighting: SourceHighlighting) {
+    this.sourceHighlighting = sourceHighlighting;
+    
     this.taskMap = {
       'CU Generation': executeCUGenTask,
       'Dependence Profiling': executeDepProfTask,
@@ -26,32 +35,48 @@ export class DiscoPoPViewProvider implements vscode.WebviewViewProvider {
     };
     this.folderPath = vscode.workspace.workspaceFolders![0].uri.path; //TODO: think about how to properly do this    ;
     this.buildPath = `${vscode.workspace.getConfiguration("discopopvsc").get("build_path")}/`;
-
-
+    this.discopopPath = `${vscode.workspace.getConfiguration("discopopvsc").get("path")}`;
+    
+    //TODO: commented cuz error
+    //this.sourceHighlighting.setView(this);
+    
     this.init = false;
   }
 
   onReceiveMessage(message: any) {
     switch (message.command) {
       case 'createFileMapping':
-        executeFileMappingTask(this);
+        //TODO: commented cuz error
+    //executeFileMappingTask(this);
         this.webview!.webview.html = this.webview!.webview.html;
         break;
       case 'startTasks':
-        message.tasks.forEach((task: any) => {
-          if(task.selected){
-            this.taskMap[task.name](this, message.files.filter((m: any) => m.selected).map((m : any) => m.file.path));
-          }
-        });
+        {
+          const tempfiles = message.files.filter((m: any) => m.selected).map((m: any) => m.file.path);
+          message.tasks.forEach((task: any) => {
+            if (task.selected) {
+              this.taskMap[task.name](this, tempfiles);
+            }
+          });
+          this.lastFiles = tempfiles;
+        }
         break;
+      case 'identifyPatterns':
+        {
+          const tempfiles = message.files.filter((m: any) => m.selected).map((m: any) => m.file.path);
+          //TODO: commented cuz error
+    //executePatIdTask(this, tempfiles);
+          this.lastFiles = tempfiles;
+          break;
+        }
       default:
         break;
     };
   }
 
-  drawWebView(){
+  drawWebView() {
     console.log("DRAWING");
-    this.webview!.webview.html =  `<!DOCTYPE html>
+    this.webview!.webview.html = `<!DOCTYPE html>
     <html lang="en">
     <head>
     <meta charset="UTF-8">
@@ -67,15 +92,18 @@ export class DiscoPoPViewProvider implements vscode.WebviewViewProvider {
     function createFileMapping(){
       vscode.postMessage({command: 'createFileMapping'});
     }
+    function identifyPatterns(){
+      vscode.postMessage({command: 'identifyPatterns', files: Array.from(document.querySelectorAll(".discopop_file")).map(el => { return {'file': { 'name': el.children[1].innerText, 'path': el.children[1].attributes.path.value }, 'selected': el.children[0].checked}}), tasks: Array.from(document.querySelectorAll(".discopop_task")).map(el => { return { 'name': el.children[1].innerText, 'selected': el.children[0].checked}})});
+    }
     </script>
     <span class="discopop-title">Files</span>
     <div id="discopopfiles">
     ${getFiles(this.folderPath).map((filepath) => {
-        return `<div class="discopop_file" title="${nameFromPath(this.folderPath, filepath)}">
+      return `<div class="discopop_file" title="${nameFromPath(this.folderPath, filepath)}">
       <input type="checkbox" tabindex="-1">
       <span path="${filepath}"> ${nameFromPath(this.folderPath, filepath)} </span>
       </div>`;
-      }).join("")}
+    }).join("")}
     </div>
     <button type="button" onclick="createFileMapping()">Create File Mapping</button>
     <br><hr><br>
@@ -97,6 +125,7 @@ export class DiscoPoPViewProvider implements vscode.WebviewViewProvider {
     <br>
     <div>
       <button type="button" onclick="startTasks()">Start Tasks</button>
+      <button type="button" onclick="identifyPatterns()">Identify Patterns</button>
     </div>
   
     </body>