+++ src/extension.ts
@@ -2,22 +2,39 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 
-import { DiscoPoPViewProvider } from './discopop_webview_provider';
+import { DiscoPoPViewProvider } from './newdiscopop_webview_provider';
 import { executeFileMappingTask } from './tasks/task_file_mapping';
 import { executeCUGenTask } from './tasks/task_cu_gen';
 import { executeDepProfTask } from './tasks/task_dep_prof';
 import { executeRedOpTask } from './tasks/task_red_op';
 
 import { SourceHighlighting } from './misc/source_highlighting';
+import { CodeActions } from './misc/code_actions';
+import { createFolderIfNotExist } from './misc/iomanip';
 
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 
+	createFolderIfNotExist(vscode.workspace.workspaceFolders![0].uri.path + '/discopop-tmp');
+
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
 	console.log('Congratulations, your extension "discopopvsc" is now active!');
+	let sourceHighlighting = new SourceHighlighting(context);
+	context.subscriptions.push(
+		vscode.languages.registerCodeActionsProvider('*', new CodeActions(sourceHighlighting), {
+			providedCodeActionKinds: CodeActions.providedCodeActionKinds
+		}));
+	const diagnostics = vscode.languages.createDiagnosticCollection("emoji");
+	context.subscriptions.push(diagnostics);
+
+	sourceHighlighting.subscribeToDocumentChanges(context, diagnostics);
+
+	// decorator
+
+
 
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
@@ -34,7 +51,7 @@ export function activate(context: vscode.ExtensionContext) {
 	//const folderPath = vscode.workspace.workspaceFolders![0].uri.path; //TODO: think about how to properly do this
 
 
-	const viewProvider: DiscoPoPViewProvider = new DiscoPoPViewProvider();
+	const viewProvider: DiscoPoPViewProvider = new DiscoPoPViewProvider(sourceHighlighting);
 
 	//let filemapping = fs.readFileSync(folderPath + '/FileMapping.txt', 'utf8').split('\n').filter((el) => el !== '');
 	//filemapping = filemapping.map(str => str.trim().substr(2));
@@ -73,14 +90,11 @@ export function activate(context: vscode.ExtensionContext) {
 
 	});
 
-	
+	const fileMappingInterval = setInterval(() => {
+		executeFileMappingTask(viewProvider);
+	}, 10000);
 
-	const emojiDiagnostics = vscode.languages.createDiagnosticCollection("emoji");
-	context.subscriptions.push(emojiDiagnostics);
 
-	let sourceHighlighting = new SourceHighlighting();
-	sourceHighlighting.subscribeToDocumentChanges(context, emojiDiagnostics);
-	
 }